Allow override with recalling chords
This patch modifies the behaviour of ^ when `remember-chords` is active
so that if it is followed by an argument in braces then that argument is
used to replace the chord that would otherwise be inserted.  The
sequencing is kept, so that a chord from the list is skipped (it would
be better if this was controlled by an option, I guess).  An empty
argument is the same as no argument.